chore: moved elevated button to local widget dir
